Dthompson/feat update stuff (#410)
* fix(build): save asm files if config is set

* feat(match): wildcard

* Update src/Parser/AST/Statements/Match.cpp


* Update src/Parser/AST/Statements/Match.cpp


---------

+++ src/main.cpp
@@ -637,7 +637,7 @@ ModuleResult compileModule(const std::string &mod, const cfg::Config &config,
       if (system(cmd.c_str()) != 0) {
         return {objPath, false};
       }
-      fs::remove(asmPath);
+      if (!config.asm_) fs::remove(asmPath);
     } else {
       return {objPath, false};
     }
